requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
pandas==1.3.3
scikit-learn==0.24.2
matplotlib==3.4.3
numpy==1.21.2